Version: 0.1.141 Branch: master Date: 02.07.2017 Author: AlexMerlyan
Working: complimentary and category controller, new DTO
Changes: complimentary and category controller, new DTO
Next:

Comment:

+++ src/main/java/projectpackage/controllers/CategoryController.java
@@ -7,12 +7,14 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import projectpackage.dto.CategoryDTO;
 import projectpackage.dto.JacksonMappingMarker;
 import projectpackage.model.auth.User;
 import projectpackage.model.orders.Category;
 import projectpackage.dto.IUDAnswer;
 import projectpackage.service.MessageBook;
 import projectpackage.service.orderservice.CategoryService;
+import projectpackage.service.support.ServiceUtils;
 
 import javax.cache.annotation.CacheRemoveAll;
 import javax.cache.annotation.CacheResult;
@@ -32,6 +34,9 @@ public class CategoryController {
 
     @Autowired
     CategoryService categoryService;
+
+    @Autowired
+    ServiceUtils serviceUtils;
 
     //Get Category List
     @ResponseStatus(HttpStatus.OK)
@@ -64,8 +69,6 @@ public class CategoryController {
         Resource<Category> resource = new Resource<>(category);
         HttpStatus status;
         if (null!=category){
-            if (thisUser.getRole().getRoleName().equals("ADMIN")) resource.add(linkTo(methodOn(CategoryController.class).deleteCategory(category.getObjectId())).withRel("delete"));
-            resource.add(linkTo(methodOn(CategoryController.class).updateCategory(category.getObjectId(), category)).withRel("update"));
             status = HttpStatus.ACCEPTED;
         } else {
             status = HttpStatus.BAD_REQUEST;
@@ -77,8 +80,16 @@ public class CategoryController {
     //Create category, fetch into database
     @CacheRemoveAll(cacheName = "categoryList")
     @RequestMapping(method = RequestMethod.POST, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE}, consumes = {MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<IUDAnswer> createCategory(@RequestBody Category newCategory){
-        IUDAnswer result = categoryService.insertCategory(newCategory);
+    public ResponseEntity<IUDAnswer> createCategory(@RequestBody CategoryDTO categoryDTO, HttpServletRequest request){
+        User user = (User) request.getSession().getAttribute("USER");
+        IUDAnswer iudAnswer = serviceUtils.checkSessionAdminAndData(user, categoryDTO);
+        if (!iudAnswer.isSuccessful()) {
+            return new ResponseEntity<IUDAnswer>(iudAnswer, HttpStatus.BAD_REQUEST);
+        }
+        Category category = new Category();
+        category.setCategoryTitle(categoryDTO.getCategoryTitle());
+        category.setCategoryPrice(categoryDTO.getCategoryPrice());
+        IUDAnswer result = categoryService.insertCategory(category);
         HttpStatus status;
         if (result.isSuccessful()) {
             status = HttpStatus.CREATED;
@@ -90,11 +101,16 @@ public class CategoryController {
     //Update category method
     @CacheRemoveAll(cacheName = "categoryList")
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE}, consumes = {MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<IUDAnswer> updateCategory(@PathVariable("id") Integer id, @RequestBody Category changedCategory){
-        if (!id.equals(changedCategory.getObjectId())){
-            return new ResponseEntity<IUDAnswer>(new IUDAnswer(id, false, MessageBook.WRONG_UPDATE_ID), HttpStatus.NOT_ACCEPTABLE);
+    public ResponseEntity<IUDAnswer> updateCategory(@PathVariable("id") Integer id, @RequestBody CategoryDTO categoryDTO, HttpServletRequest request){
+        User user = (User) request.getSession().getAttribute("USER");
+        IUDAnswer iudAnswer = serviceUtils.checkSessionAdminAndData(user, categoryDTO, id);
+        if (!iudAnswer.isSuccessful()) {
+            return new ResponseEntity<IUDAnswer>(iudAnswer, HttpStatus.BAD_REQUEST);
         }
-        IUDAnswer result = categoryService.updateCategory(id, changedCategory);
+        Category category = new Category();
+        category.setCategoryTitle(categoryDTO.getCategoryTitle());
+        category.setCategoryPrice(categoryDTO.getCategoryPrice());
+        IUDAnswer result = categoryService.updateCategory(id, category);
         HttpStatus status;
         if (result.isSuccessful()) {
             status = HttpStatus.ACCEPTED;
@@ -106,7 +122,12 @@ public class CategoryController {
     //Delete category method
     @CacheRemoveAll(cacheName = "categoryList")
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<IUDAnswer> deleteCategory(@PathVariable("id") Integer id){
+    public ResponseEntity<IUDAnswer> deleteCategory(@PathVariable("id") Integer id, HttpServletRequest request){
+        User user = (User) request.getSession().getAttribute("USER");
+        IUDAnswer iudAnswer = serviceUtils.checkDeleteForAdmin(user, id);
+        if (!iudAnswer.isSuccessful()) {
+            return new ResponseEntity<IUDAnswer>(iudAnswer, HttpStatus.BAD_REQUEST);
+        }
         IUDAnswer result = categoryService.deleteCategory(id);
         HttpStatus status;
         if (result.isSuccessful()) {
